+++ apps/io-wallet-user-func/src/infra/attestation-service/android/index.ts
@@ -1,3 +1,4 @@
+import { getCrlFromUrl } from "@/certificates";
 import { parse } from "@pagopa/handler-kit";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { X509Certificate } from "crypto";
@@ -10,7 +11,7 @@ import * as S from "fp-ts/lib/string";
 import { JwkPublicKey } from "io-wallet-common/jwk";
 
 import { ValidatedAttestation } from "../../attestation-service";
-import { AndroidAttestationError } from "../errors";
+import { AndroidAssertionError, AndroidAttestationError } from "../errors";
 import { GoogleAppCredentials, verifyAssertion } from "./assertion";
 import { verifyAttestation } from "./attestation";
 
@@ -35,10 +36,7 @@ export const validateAndroidAttestation = (
         E.tryCatch(
           () => new X509Certificate(cert),
           () =>
-            new AndroidAttestationError(
-              `Unable to decode X509 certificate`,
-              {},
-            ),
+            new AndroidAttestationError(`Unable to decode X509 certificate`),
         ),
       ),
     ),
@@ -46,17 +44,26 @@ export const validateAndroidAttestation = (
     TE.fromEither,
     TE.chain((x509Chain) =>
       pipe(
-        TE.tryCatch(
-          () =>
-            verifyAttestation({
-              androidCrlUrl,
-              bundleIdentifiers,
-              challenge: nonce,
-              googlePublicKey,
-              httpRequestTimeout,
-              x509Chain,
-            }),
-          E.toError,
+        getCrlFromUrl(androidCrlUrl, httpRequestTimeout),
+        TE.chain((attestationCrl) =>
+          TE.tryCatch(
+            () =>
+              verifyAttestation({
+                attestationCrl,
+                bundleIdentifiers,
+                challenge: nonce,
+                googlePublicKey,
+                x509Chain,
+              }),
+            E.toError,
+          ),
+        ),
+        TE.chain((attestationValidationResult) =>
+          attestationValidationResult.success
+            ? TE.right(attestationValidationResult)
+            : TE.left(
+                new AndroidAttestationError(attestationValidationResult.reason),
+              ),
         ),
         TE.chainW(({ deviceDetails, hardwareKey }) =>
           pipe(
@@ -89,8 +96,8 @@ export const validateAndroidAssertion = (
     E.chain(J.parse),
     E.mapLeft(
       () =>
-        new Error(
-          "[Android Assertion] Unable to parse Google App Credentials string",
+        new AndroidAssertionError(
+          "Unable to parse Google App Credentials string",
         ),
     ),
     E.chainW(parse(GoogleAppCredentials, "Invalid Google App Credentials")),
@@ -111,5 +118,10 @@ export const validateAndroidAssertion = (
           }),
         E.toError,
       ),
+    ),
+    TE.chain((assertionValidationResult) =>
+      assertionValidationResult.success
+        ? TE.right(undefined)
+        : TE.left(new AndroidAssertionError(assertionValidationResult.reason)),
     ),
   );